% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_answer.R
\name{add_answer}
\alias{add_answer}
\title{Add an answer to your pub quiz slides}
\usage{
add_answer(x)
}
\arguments{
\item{x}{ID from a database}
}
\description{
This function knits a child Rmd document with an answer to a question
for use in a pub quiz template.
Answers are typically presented at the end of the pub quiz slides.
}
\examples{
add_answer()
}
\keyword{question}
