% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_question.R
\name{add_question}
\alias{add_question}
\title{Add a question to your pub quiz slides}
\usage{
add_question(x, db = data(pubquizr_db))
}
\arguments{
\item{x}{ID from a database}
}
\description{
This function knits a child Rmd document with a question for use in a pub quiz template.
}
\examples{
add_question()
}
\keyword{question}
