% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_question_base.R
\name{prepare_question_base}
\alias{prepare_question_base}
\title{FIlter a set of question based on your needs}
\usage{
prepare_question_base(
  db,
  n = 10,
  include_text = TRUE,
  include_audio = TRUE,
  include_image = FALSE
)
}
\arguments{
\item{db}{database object}

\item{n}{number of questions to keep}

\item{include_text}{Can the set contain questions with text?}

\item{include_audio}{Can the set contain questions with audio?}

\item{include_image}{Can the set contain questions with images?}
}
\description{
Filters a set of questions for further use in a pub quiz slide presentation.
The function expects that the supplied data frame contains variables with the following names: description, audio, question_image
}
\examples{
prepare_question_base()
}
\keyword{prep}
