% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pubquizr_db}
\alias{pubquizr_db}
\title{Pub quiz example dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 8 columns.
}
\usage{
data(pubquizr_db)
}
\description{
Example pub quiz questions to generate a slide deck.
}
\examples{
data(pubquizr_db)
}
\keyword{datasets}
